fix and load migrations

+++ publishables/migrations/2019_08_24_042849_create_product_prdcats_table.php
@@ -15,8 +15,8 @@ class CreateProductPrdcatsTable extends Migration
     {
         Schema::create('product_prdcats', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('product_id',false,true)->default(0);
-            $table->integer('category_id',false,true)->default(0);
+            $table->bigInteger('product_id',false,true)->default(0);
+            $table->bigInteger('category_id',false,true)->default(0);
             $table->timestamps();
 
             $table->unique(['product_id','category_id']);
